+++ SharpMUSH.Implementation/Commands/MoreCommands.cs
@@ -1,11 +1,12 @@
 ﻿using OneOf.Types;
+using SharpMUSH.Library.Attributes;
 using SharpMUSH.Library.DiscriminatedUnions;
 using SharpMUSH.Library.ParserInterfaces;
-using CB = SharpMUSH.Implementation.Definitions.CommandBehavior;
+using CB = SharpMUSH.Library.Definitions.CommandBehavior;
 
 namespace SharpMUSH.Implementation.Commands;
 
-public static partial class Commands
+public partial class Commands
 {
 	[SharpCommand(Name = "@CLOCK", Switches = ["JOIN", "SPEAK", "MOD", "SEE", "HIDE"], Behavior = CB.Default | CB.EqSplit,
 		MinArgs = 0, MaxArgs = 0)]
