+++ SharpMUSH.Implementation/Commands/SocketCommands.cs
@@ -1,18 +1,19 @@
 ﻿using OneOf.Types;
-using Serilog;
 using SharpMUSH.Library.DiscriminatedUnions;
 using SharpMUSH.Library.Models;
 using SharpMUSH.Library.ParserInterfaces;
+using SharpMUSH.Library.Definitions;
 using SharpMUSH.Library.Queries.Database;
 using System.Text.RegularExpressions;
+using SharpMUSH.Library.Attributes;
 
 namespace SharpMUSH.Implementation.Commands;
 
-public static partial class Commands
+public partial class Commands
 {
 	private static readonly Regex ConnectionPatternRegex = ConnectionPattern();
 
-	[SharpCommand(Name = "WHO", Behavior = Definitions.CommandBehavior.SOCKET | Definitions.CommandBehavior.NoParse, MinArgs = 0, MaxArgs = 1)]
+	[SharpCommand(Name = "WHO", Behavior = CommandBehavior.SOCKET | CommandBehavior.NoParse, MinArgs = 0, MaxArgs = 1)]
 	public static async ValueTask<Option<CallState>> Who(IMUSHCodeParser parser, SharpCommandAttribute _2)
 	{
 		// TODO: Get All needs to do a Permission Check for the user.
@@ -46,7 +47,7 @@ public static partial class Commands
 	/// connect PersonWithoutAPassword
 	/// connect "person without a password"
 	/// </example>
-	[SharpCommand(Name = "CONNECT", Behavior = Definitions.CommandBehavior.SOCKET | Definitions.CommandBehavior.NoParse, MinArgs = 1, MaxArgs = 2)]
+	[SharpCommand(Name = "CONNECT", Behavior = CommandBehavior.SOCKET | CommandBehavior.NoParse, MinArgs = 1, MaxArgs = 2)]
 	public static async ValueTask<Option<CallState>> Connect(IMUSHCodeParser parser, SharpCommandAttribute _2)
 	{
 		// Early HUH if already logged in.
@@ -99,7 +100,7 @@ public static partial class Commands
 		return new None();
 	}
 
-	[SharpCommand(Name = "QUIT", Behavior = Definitions.CommandBehavior.SOCKET | Definitions.CommandBehavior.NoParse, MinArgs = 0, MaxArgs = 0)]
+	[SharpCommand(Name = "QUIT", Behavior = CommandBehavior.SOCKET | CommandBehavior.NoParse, MinArgs = 0, MaxArgs = 0)]
 	public static ValueTask<Option<CallState>> Quit(IMUSHCodeParser parser, SharpCommandAttribute _2)
 	{
 		// TODO: Display Disconnect Banner.
