+++ SharpMUSH.Implementation/Visitors/SharpMUSHParserVisitor.cs
@@ -1,4 +1,5 @@
-﻿using System.Runtime.CompilerServices;
+﻿using System.Collections.Immutable;
+using System.Runtime.CompilerServices;
 using Antlr4.Runtime.Misc;
 using Antlr4.Runtime.Tree;
 using Microsoft.Extensions.Logging;
@@ -8,7 +9,11 @@ using SharpMUSH.Implementation.Functions;
 using SharpMUSH.Library.Attributes;
 using SharpMUSH.Library.Definitions;
 using SharpMUSH.Library.DiscriminatedUnions;
+using SharpMUSH.Library.Extensions;
+using SharpMUSH.Library.Models;
 using SharpMUSH.Library.ParserInterfaces;
+using SharpMUSH.Library.Queries.Database;
+using SharpMUSH.Library.Services;
 using static SharpMUSHParser;
 
 namespace SharpMUSH.Implementation.Visitors;
@@ -85,14 +90,14 @@ public class SharpMUSHParserVisitor(ILogger logger, IMUSHCodeParser parser, MStr
 	/// TODO: Optimization needed. We should at least grab the in-built ones at startup.
 	/// TODO: Move this to a Library Service implementation.
 	/// </summary>
-	/// <param name="logger"></param>
+	/// <param name="newLogger"></param>
 	/// <param name="name">Function Name</param>
-	/// <param name="source">The source MarkupString</param>
+	/// <param name="src">The source MarkupString</param>
 	/// <param name="context">Function Context for Depth</param>
 	/// <param name="args">Arguments</param>
 	/// <param name="visitor"></param>
 	/// <returns>The resulting CallState.</returns>
-	public async ValueTask<CallState> CallFunction(ILogger logger, string name, MString source,
+	public async ValueTask<CallState> CallFunction(ILogger newLogger, string name, MString src,
 		FunctionContext context, EvaluationStringContext[] args, SharpMUSHParserVisitor visitor)
 	{
 		try
@@ -178,7 +183,7 @@ public class SharpMUSHParserVisitor(ILogger logger, IMUSHCodeParser parser, MStr
 			else if (attribute.Flags.HasFlag(FunctionFlags.NoParse) && attribute.MaxArgs == 1)
 			{
 				return new CallState(
-					MModule.substring(context.Start.StartIndex, context.Stop.StopIndex - context.Start.StartIndex + 1, source),
+					MModule.substring(context.Start.StartIndex, context.Stop.StopIndex - context.Start.StartIndex + 1, src),
 					contextDepth,
 					null,
 					async () => (await visitor.VisitChildren(context) ?? CallState.Empty with { Depth = context.Depth() })
@@ -188,9 +193,9 @@ public class SharpMUSHParserVisitor(ILogger logger, IMUSHCodeParser parser, MStr
 			{
 				refinedArguments = args.Select(x => new CallState(stripAnsi
 							? MModule.plainText2(MModule.substring(x.Start.StartIndex,
-								context.Stop?.StopIndex is null ? 0 : (x.Stop.StopIndex - x.Start.StartIndex + 1), source))
+								context.Stop?.StopIndex is null ? 0 : (x.Stop.StopIndex - x.Start.StartIndex + 1), src))
 							: MModule.substring(x.Start.StartIndex,
-								context.Stop?.StopIndex is null ? 0 : (x.Stop.StopIndex - x.Start.StartIndex + 1), source),
+								context.Stop?.StopIndex is null ? 0 : (x.Stop.StopIndex - x.Start.StartIndex + 1), src),
 						x.Depth(), null,
 						async () => stripAnsi
 							? (await visitor.VisitChildren(x))!.Message
@@ -200,13 +205,13 @@ public class SharpMUSHParserVisitor(ILogger logger, IMUSHCodeParser parser, MStr
 			}
 
 			if (attribute.Flags.HasFlag(FunctionFlags.DecimalsOnly) && refinedArguments.Any(a =>
-				    !decimal.TryParse(MModule.plainText(a?.Message ?? MModule.empty()), out _)))
+				    !decimal.TryParse(MModule.plainText(a.Message ?? MModule.empty()), out _)))
 			{
 				return new CallState(attribute.MaxArgs > 1 ? Errors.ErrorNumbers : Errors.ErrorNumber);
 			}
 
 			if (attribute.Flags.HasFlag(FunctionFlags.IntegersOnly) && refinedArguments.Any(a =>
-				    !int.TryParse(MModule.plainText(a?.Message ?? MModule.empty()), out _)))
+				    !int.TryParse(MModule.plainText(a.Message ?? MModule.empty()), out _)))
 			{
 				return new CallState(attribute.MaxArgs > 1 ? Errors.ErrorIntegers : Errors.ErrorInteger);
 			}
@@ -247,9 +252,449 @@ public class SharpMUSHParserVisitor(ILogger logger, IMUSHCodeParser parser, MStr
 		if (!parser.FunctionLibrary.TryGetValue(name, out var result) && result.IsSystem)
 			return new None();
 
-		return (result.LibraryInformation.Attribute, new Func<IMUSHCodeParser, ValueTask<CallState>>
-			(p => (ValueTask<CallState>)result.LibraryInformation.Function.Method.Invoke(null, [p, result.LibraryInformation.Attribute])!));
+		return (result.LibraryInformation.Attribute, 
+			p => (ValueTask<CallState>)result.LibraryInformation.Function.Method.Invoke(null, [p, result.LibraryInformation.Attribute])!);
 	}
+	
+	
+	/// <summary>
+	/// Evaluates the command, with the parser info given.
+	/// </summary>
+	/// <remarks>
+	/// Call State is expected to be empty on return. 
+	/// But if one wanted to implement a @pipe command that can pass a result from say, a @dig command, 
+	/// there would be a need for some way of passing on secondary data.
+	/// </remarks>
+	/// <param name="logger">Logger</param>
+	/// <param name="source">Original string</param>
+	/// <param name="context">Command Context</param>
+	/// <param name="visitChildren">Parser function to visit children.</param>
+	/// <returns>An empty Call State</returns>
+	public async ValueTask<Option<CallState>> EvaluateCommands(ILogger newLogger, 
+		MString src,
+		CommandContext context,
+		Func<IRuleNode, ValueTask<CallState?>> visitChildren)
+	{
+		try
+		{
+			var firstCommandMatch = context.evaluationString();
+
+			if (firstCommandMatch?.SourceInterval.Length is null or 0)
+				return new None();
+
+			var command = firstCommandMatch.GetText();
+
+			var spaceIndex = command.AsSpan().IndexOf(' ');
+			if (spaceIndex != -1)
+			{
+				command = command[..spaceIndex];
+			}
+
+			if (parser.CurrentState.Handle is not null && command == "IDLE")
+			{
+				parser.ConnectionService.Update(parser.CurrentState.Handle.Value, "LastConnectionSignal",
+					DateTimeOffset.UtcNow.ToUnixTimeMilliseconds().ToString());
+			}
+
+			// Step 1: Check if it's a SOCKET command
+			// TODO: Optimize
+			var socketCommandPattern = parser.CommandLibrary.Where(x
+				=> parser.CurrentState.Handle is not null
+				   && x.Value.IsSystem
+				   && x.Key.Equals(command, StringComparison.CurrentCultureIgnoreCase)
+				   && x.Value.LibraryInformation.Attribute.Behavior.HasFlag(CommandBehavior.SOCKET)).ToList();
+
+			if (socketCommandPattern.Any() &&
+			    parser.CommandLibrary.TryGetValue(command.ToUpper(), out var librarySocketCommandDefinition))
+			{
+				return await HandleSocketCommandPattern(parser, src, context, command, socketCommandPattern,
+					librarySocketCommandDefinition.LibraryInformation);
+			}
+
+			if (parser.CurrentState.Executor is null && parser.CurrentState.Handle is not null)
+			{
+				await parser.NotifyService.Notify(parser.CurrentState.Handle.Value, "No such command available at login.");
+				return new None();
+			}
+
+			// Step 2a: Check for the channel single-token command.
+
+			// TODO: Better channel name matching within the channel helper.
+			if (command[..1] == parser.Configuration.CurrentValue.Chat.ChatTokenAlias.ToString())
+			{
+				var channels = await parser.Mediator.Send(new GetChannelListQuery());
+				var check = command[1..];
+
+				var channel = channels.FirstOrDefault(x =>
+					x.Name.ToPlainText().StartsWith(check, StringComparison.CurrentCultureIgnoreCase));
+
+				if (channel is not null && !context.evaluationString().IsEmpty)
+				{
+					return await HandleChannelCommand(parser, channel, context, src);
+				}
+			}
+
+			// Step 2b: Check for a single-token command
+			// TODO: Optimize
+			var singleTokenCommandPattern = parser.CommandLibrary.Where(x
+				=> x.Key.Equals(command[..1], StringComparison.CurrentCultureIgnoreCase) 
+				   && x.Value.IsSystem 
+				   && x.Value.LibraryInformation.Attribute.Behavior.HasFlag(CommandBehavior.SingleToken)).ToList();
+
+			if (singleTokenCommandPattern.Count != 0)
+			{
+				return await HandleSingleTokenCommandPattern(parser, src, context, command, singleTokenCommandPattern);
+			}
+
+			var executorObject = (await parser.CurrentState.ExecutorObject(parser.Mediator)).WithoutNone();
+			// Step 3: Check exit Aliases
+			if (executorObject.IsContent)
+			{
+				var locate = await parser.LocateService.Locate(
+					parser,
+					executorObject,
+					executorObject,
+					command,
+					LocateFlags.ExitsInTheRoomOfLooker
+					| LocateFlags.EnglishStyleMatching
+					| LocateFlags.ExitsPreference
+					| LocateFlags.OnlyMatchTypePreference
+					| LocateFlags.FailIfNotPreferred);
+
+				if (locate.IsExit)
+				{
+					var exit = locate.AsExit;
+					return await HandleGoCommandPattern(parser, exit);
+				}
+			}
+
+			// Step 4: Check if we are setting an attribute: &... -- we're just treating this as a Single Token Command for now.
+			// Who would rely on a room alias being & anyway?
+			// Step 5: Check @COMMAND in command library
+
+			// TODO: Optimize
+			// TODO: Get the Switches and send them along as a list of items!
+			var slashIndex = command.AsSpan().IndexOf('/');
+			var rootCommand =
+				command[..(slashIndex > -1 ? slashIndex : command.Length)];
+			var swtch = command[(slashIndex > -1 ? slashIndex : command.Length)..];
+			var switches = swtch.Split('/').Where(s => !string.IsNullOrWhiteSpace(s));
+
+			if (parser.CommandLibrary.TryGetValue(rootCommand.ToUpper(), out var libraryCommandDefinition)
+			    && !rootCommand.Equals("HUH_COMMAND", StringComparison.CurrentCultureIgnoreCase))
+			{
+				return await HandleInternalCommandPattern(parser, src, context, rootCommand, switches,
+					libraryCommandDefinition.LibraryInformation);
+			}
+
+			// Step 6: Check @attribute setting
+			// Step 7: Enter Aliases
+			// Step 8: Leave Aliases
+
+			// Step 9: User Defined Commands nearby
+			// -- This is going to be a very important place to Cache the commands.
+			// A caching strategy is going to be reliant on the Attribute Service.
+			// Optimistic that the command still exists, until we try and it no longer does?
+			// What's the best way to retrieve the Regex or Wildcard pattern and transform it? 
+			// It needs to take an area to search in. So this is definitely its own service.
+			var nearbyObjects = await parser.Mediator.Send(new GetNearbyObjectsQuery(executorObject.Object().DBRef));
+
+			var userDefinedCommandMatches = await parser.CommandDiscoveryService.MatchUserDefinedCommand(
+				parser,
+				nearbyObjects,
+				src);
+
+			if (userDefinedCommandMatches.IsSome())
+			{
+				return await HandleUserDefinedCommand(parser, userDefinedCommandMatches.AsValue());
+			}
+
+			// Step 10: Zone Exit Name and Aliases
+			// Step 11: Zone Master User Defined Commands
+			// Step 12: User Defined commands on the location itself.
+			if (executorObject.IsContent)
+			{
+				var userDefinedCommandMatchesOnLocation = await parser.CommandDiscoveryService.MatchUserDefinedCommand(
+					parser,
+					[(await executorObject.AsContent.Location()).WithExitOption()],
+					src);
+
+				if (userDefinedCommandMatchesOnLocation.IsSome())
+				{
+					return await HandleUserDefinedCommand(parser, userDefinedCommandMatchesOnLocation.AsValue());
+				}
+			}
+
+			// Step 13: User defined commands on the player's personal zone.
+			// Step 14: Global Exits
+			// Step 15: Global User-defined commands
+			var goConfig = parser.Configuration.CurrentValue.Database.MasterRoom;
+			var maybeGlobalObject = await parser.Mediator.Send(new GetObjectNodeQuery(new DBRef(Convert.ToInt32(goConfig))));
+			var globalObject = maybeGlobalObject.Known();
+			var globalObjectContent = (await globalObject.AsContainer.Content(parser))
+				.Select(x => x.WithRoomOption());
+
+			var userDefinedCommandMatchesOnGlobal = await parser.CommandDiscoveryService.MatchUserDefinedCommand(
+				parser,
+				[globalObject, .. globalObjectContent],
+				src);
+
+			if (userDefinedCommandMatchesOnGlobal.IsSome())
+			{
+				return await HandleUserDefinedCommand(parser, userDefinedCommandMatchesOnGlobal.AsValue());
+			}
+
+			// Step 16: HUH_COMMAND is run
+			var newParser = parser.Push(parser.CurrentState with
+			{
+				Command = "HUH_COMMAND",
+				Arguments = [],
+				Function = null
+			});
+
+			var huhCommand = await parser.CommandLibrary["HUH_COMMAND"].LibraryInformation.Command.Invoke(newParser);
+
+			return huhCommand;
+		}
+		catch (Exception ex)
+		{
+			logger.LogError(ex, nameof(EvaluateCommands));
+			return CallState.Empty;
+		}
+	}
+
+	private async Task<Option<CallState>> HandleChannelCommand(IMUSHCodeParser parser, SharpChannel channel,
+		CommandContext context, MString src)
+	{
+		var rest = MModule.substring(
+			context.evaluationString().Start.StartIndex,
+			context.evaluationString().Stop.StopIndex - context.evaluationString().Start.StartIndex + 1,
+			src);
+
+		var chatParser = parser.Push(parser.CurrentState with
+		{
+			Command = "@CHAT",
+			Arguments = new Dictionary<string, CallState>
+			{
+				{ "0", new CallState(channel.Name) },
+				{ "1", new CallState(rest) }
+			}
+		});
+
+		return await parser.CommandLibrary["@CHAT"].LibraryInformation.Command.Invoke(chatParser);
+	}
+
+	private async Task<Option<CallState>> HandleUserDefinedCommand(
+		IMUSHCodeParser parser,
+		IEnumerable<(AnySharpObject Obj, SharpAttribute Attr, Dictionary<string, CallState> Arguments)> matches)
+	{
+		// Step 1: Validate if the command can be evaluated (locks)
+		foreach (var (obj, attr, arguments) in matches)
+		{
+			var newParser = parser.Push(parser.CurrentState with
+			{
+				CurrentEvaluation = new DBAttribute(obj.Object().DBRef, attr.Name),
+				Arguments = arguments,
+				Function = null,
+				Executor = obj.Object().DBRef
+			});
+
+			await newParser.CommandListParse(MModule.substring(
+				attr.CommandListIndex!.Value,
+				MModule.getLength(attr.Value) - attr.CommandListIndex!.Value,
+				attr.Value));
+		}
+
+		return CallState.Empty;
+	}
+
+	private async ValueTask<Option<CallState>> HandleGoCommandPattern(IMUSHCodeParser parser, SharpExit exit)
+	{
+		var newParser = parser.Push(parser.CurrentState with
+		{
+			Command = "GOTO",
+			Arguments = new Dictionary<string, CallState>
+			{
+				{ "0", new CallState(exit.Object.DBRef.ToString(), 0) }
+			},
+			Function = null
+		});
+
+		return await parser.CommandLibrary["GOTO"].LibraryInformation.Command.Invoke(newParser);
+	}
+
+	private static async ValueTask<Option<CallState>> HandleInternalCommandPattern(IMUSHCodeParser parser, MString src,
+		CommandContext context, string rootCommand, IEnumerable<string> switches,
+		CommandDefinition libraryCommandDefinition)
+	{
+		var arguments = await ArgumentSplit(parser, src, context, libraryCommandDefinition);
+
+		var newParser = parser.Push(parser.CurrentState with
+		{
+			Command = rootCommand,
+			Switches = switches.Select(x => x.ToUpper()),
+			Arguments = arguments
+				.Select((value, i) => new KeyValuePair<string, CallState>(i.ToString(), value))
+				.ToDictionary(),
+			Function = null
+		});
+
+		return await libraryCommandDefinition.Command.Invoke(newParser);
+	}
+
+	private static async ValueTask<Option<CallState>> HandleSocketCommandPattern(IMUSHCodeParser parser, MString src,
+		CommandContext context, string command,
+		IEnumerable<KeyValuePair<string, (CommandDefinition LibraryInformation,bool IsSystem)>> socketCommandPattern,
+		(SharpCommandAttribute Attribute, Func<IMUSHCodeParser, ValueTask<Option<CallState>>> Function)
+			librarySocketCommandDefinition)
+	{
+		var arguments = await ArgumentSplit(parser, src, context, librarySocketCommandDefinition);
+
+		var newParser = parser.Push(parser.CurrentState with
+		{
+			Command = command,
+			Arguments = arguments.Select((value, i) => 
+					new KeyValuePair<string, CallState>(i.ToString(), value))
+				.ToDictionary(),
+			Function = null
+		});
+
+		// Run as Socket Command. 
+		return await socketCommandPattern.First().Value.LibraryInformation.Command.Invoke(newParser);
+	}
+
+	private static async ValueTask<Option<CallState>> HandleSingleTokenCommandPattern(IMUSHCodeParser parser,
+		MString src, CommandContext context, string command,
+		IEnumerable<KeyValuePair<string, (CommandDefinition LibraryInformation,bool IsSystem)>> singleTokenCommandPattern)
+	{
+		var singleRootCommand = command[..1];
+		var rest = command[1..];
+		var singleLibraryCommandDefinition = singleTokenCommandPattern.Single().Value;
+		// TODO: Should Single Commands split? - Getting errors out of this.
+		var arguments = await ArgumentSplit(parser, src, context, singleLibraryCommandDefinition.LibraryInformation);
+
+		var newParser = parser.Push(
+			parser.CurrentState with
+			{
+				Command = singleRootCommand,
+				Arguments = ImmutableDictionary<string, CallState>.Empty
+					.Add("0", new CallState(rest))
+					.AddRange(arguments.Select((value, i) => new KeyValuePair<string, CallState>((i + 1).ToString(), value)))
+					.ToDictionary(),
+				Function = null
+			}
+		);
+
+		return await singleLibraryCommandDefinition.LibraryInformation.Command.Invoke(newParser);
+	}
+
+	private static async ValueTask<List<CallState>> ArgumentSplit(IMUSHCodeParser parser, MString src,
+		CommandContext context,
+		(SharpCommandAttribute Attribute, Func<IMUSHCodeParser, ValueTask<Option<CallState>>> Function)
+			libraryCommandDefinition)
+	{
+		var argCallState = CallState.EmptyArgument;
+		var behavior = libraryCommandDefinition.Attribute.Behavior;
+
+		// Do not parse the argument splitting.
+		var newNoParseParser = parser.Push(parser.CurrentState with { ParseMode = ParseMode.NoParse });
+		var realSubtext = MModule.substring(
+			context.evaluationString().Start.StartIndex,
+			context.evaluationString().Stop.StopIndex - context.evaluationString().Start.StartIndex + 1,
+			src);
+		var spaceInContext = MModule.indexOf(realSubtext, MModule.single(" "));
+
+		// command (space) argument(s)
+		if (spaceInContext != -1)
+		{
+			var remainder =
+				MModule.substring(spaceInContext + 1, MModule.getLength(realSubtext) - spaceInContext, realSubtext);
+
+			// command arg0 = arg1,still arg 1 
+			if (behavior.HasFlag(CommandBehavior.EqSplit) && behavior.HasFlag(CommandBehavior.RSArgs))
+			{
+				argCallState = await newNoParseParser.CommandEqSplitArgsParse(remainder);
+			}
+			// command arg0 = arg1,arg2
+			else if (behavior.HasFlag(CommandBehavior.EqSplit))
+			{
+				argCallState = await newNoParseParser.CommandEqSplitParse(remainder);
+			}
+			// Command arg0,arg1,arg2,arg
+			else if (behavior.HasFlag(CommandBehavior.RSArgs))
+			{
+				argCallState = await newNoParseParser.CommandCommaArgsParse(remainder);
+			}
+			else
+			{
+				argCallState = await newNoParseParser.CommandSingleArgParse(remainder);
+			}
+		}
+
+		List<CallState> arguments = [];
+
+		var eqSplit = libraryCommandDefinition.Attribute.Behavior.HasFlag(CommandBehavior.EqSplit);
+		var noParse = libraryCommandDefinition.Attribute.Behavior.HasFlag(CommandBehavior.NoParse);
+		var noRsParse = libraryCommandDefinition.Attribute.Behavior.HasFlag(CommandBehavior.RSNoParse);
+		var nArgs = argCallState?.Arguments?.Length;
+
+		// TODO: Implement lsargs - but there are no immediate commands that need it.
+
+		if (argCallState is null)
+		{
+			return arguments;
+		}
+
+		if (eqSplit)
+		{
+			arguments.Add(noParse
+				? new CallState(argCallState.Arguments!.FirstOrDefault() ?? MModule.empty(), argCallState.Depth)
+				: (await parser.FunctionParse(argCallState.Arguments!.FirstOrDefault() ?? MModule.empty()))!);
+
+			if (nArgs < 2) return arguments;
+
+			if (noRsParse || noParse)
+			{
+				arguments.AddRange(argCallState.Arguments!
+					.Skip(1)
+					// TODO: Implement Parsed Message alt
+					.Select(x =>
+						new CallState(x,
+							argCallState.Depth,
+							null,
+							async () =>
+								(await parser.FunctionParse(x))!.Message!)));
+			}
+			else
+			{
+				foreach (var argument in argCallState.Arguments!.Skip(1))
+				{
+					// This is done to avoid allocation with ValueTask.
+					arguments.Add((await parser.FunctionParse(argument))!);
+				}
+			}
+		}
+		else
+		{
+			if (noParse)
+			{
+				arguments.AddRange(argCallState.Arguments!
+					.Select(x => new CallState(x, argCallState.Depth)));
+			}
+			else
+			{
+				foreach (var argument in argCallState.Arguments!)
+				{
+					// This is done to avoid allocation with ValueTask.
+					arguments.Add((await parser.FunctionParse(argument))!);
+				}
+			}
+		}
+
+		return arguments;
+	}
+	
+	
 	
 	public override async ValueTask<CallState?> VisitEvaluationString(
 		[NotNull] EvaluationStringContext context) => await VisitChildren(context) ?? new(
@@ -417,7 +862,7 @@ public class SharpMUSHParserVisitor(ILogger logger, IMUSHCodeParser parser, MStr
 			return await VisitChildren(context) ?? new CallState(context.GetText());
 		}
 
-		return (await Commands.Commands.EvaluateCommands(logger, parser, source, context, VisitChildren))
+		return (await EvaluateCommands(logger, source, context, VisitChildren))
 			.Match<CallState?>(
 				x => x,
 				_ => null);
