+++ SharpMUSH.Library/Definitions/LibraryProviders.cs
@@ -1,6 +1,6 @@
 ﻿using SharpMUSH.Library.Attributes;
+using SharpMUSH.Library.DiscriminatedUnions;
 using SharpMUSH.Library.ParserInterfaces;
-using SharpMUSH.Library.Services;
 
 namespace SharpMUSH.Library.Definitions;
 
@@ -14,5 +14,18 @@ public record struct FunctionDefinition(SharpFunctionAttribute Attribute, Func<I
 	public static implicit operator FunctionDefinition((SharpFunctionAttribute Attribute, Func<IMUSHCodeParser, ValueTask<CallState>> Function) value)
 	{
 		return new FunctionDefinition(value.Attribute, value.Function);
+	}
+}
+
+public record struct CommandDefinition(SharpCommandAttribute Attribute, Func<IMUSHCodeParser, ValueTask<Option<CallState>>> Command)
+{
+	public static implicit operator (SharpCommandAttribute Attribute, Func<IMUSHCodeParser, ValueTask<Option<CallState>>> Command)(CommandDefinition value)
+	{
+		return (value.Attribute, value.Command);
+	}
+
+	public static implicit operator CommandDefinition((SharpCommandAttribute Attribute, Func<IMUSHCodeParser, ValueTask<Option<CallState>>> Command) value)
+	{
+		return new CommandDefinition(value.Attribute, value.Command);
 	}
 }
